"Parcial 03052023"

+++ Leoz Mauricio/src/Models/Escritorio.java
@@ -1,36 +1,36 @@
 package Models;
 
 public class Escritorio extends Mueble{
-    private int alto;
-    private int ancho;
+    private double alto;
+    private double ancho;
 
     public Escritorio() {
     }
 
-    public Escritorio(int alto, int ancho) {
+    public Escritorio(double alto, double ancho) {
         this.alto = alto;
         this.ancho = ancho;
     }
 
-    public Escritorio(int stock, String nombre, double precio, int alto, int ancho) {
+    public Escritorio(int stock, String nombre, double precio, double alto, double ancho) {
         super(stock, nombre, precio);
         this.alto = alto;
         this.ancho = ancho;
     }
 
-    public int getAlto() {
+    public double getAlto() {
         return alto;
     }
 
-    public void setAlto(int alto) {
+    public void setAlto(double alto) {
         this.alto = alto;
     }
 
-    public int getAncho() {
+    public double getAncho() {
         return ancho;
     }
 
-    public void setAncho(int ancho) {
+    public void setAncho(double ancho) {
         this.ancho = ancho;
     }
